refactor: set chart registration done in onMounted
